refactor chains
