+++ x/observer/types/core_params.go
@@ -1,6 +1,3 @@
-//go:build TESTNET
-// +build TESTNET
-
 package types
 
 import (
@@ -13,6 +10,45 @@ import (
 func GetCoreParams() CoreParamsList {
 	params := CoreParamsList{
 		CoreParams: []*CoreParams{
+			{
+				ChainId:                     common.EthChain().ChainId,
+				ConfirmationCount:           14,
+				ZetaTokenContractAddress:    "",
+				ConnectorContractAddress:    "",
+				Erc20CustodyContractAddress: "",
+				InTxTicker:                  12,
+				OutTxTicker:                 15,
+				WatchUtxoTicker:             0,
+				GasPriceTicker:              30,
+				OutboundTxScheduleInterval:  30,
+				OutboundTxScheduleLookahead: 60,
+			},
+			{
+				ChainId:                     common.BscMainnetChain().ChainId,
+				ConfirmationCount:           14,
+				ZetaTokenContractAddress:    "",
+				ConnectorContractAddress:    "",
+				Erc20CustodyContractAddress: "",
+				InTxTicker:                  5,
+				OutTxTicker:                 15,
+				WatchUtxoTicker:             0,
+				GasPriceTicker:              30,
+				OutboundTxScheduleInterval:  30,
+				OutboundTxScheduleLookahead: 60,
+			},
+			{
+				ChainId:                     common.BtcMainnetChain().ChainId,
+				ConfirmationCount:           2,
+				ZetaTokenContractAddress:    "",
+				ConnectorContractAddress:    "",
+				Erc20CustodyContractAddress: "",
+				WatchUtxoTicker:             30,
+				InTxTicker:                  120,
+				OutTxTicker:                 60,
+				GasPriceTicker:              30,
+				OutboundTxScheduleInterval:  30,
+				OutboundTxScheduleLookahead: 60,
+			},
 			{
 				ChainId:           common.GoerliChain().ChainId,
 				ConfirmationCount: 6,
@@ -65,6 +101,19 @@ func GetCoreParams() CoreParamsList {
 				GasPriceTicker:              30,
 				OutboundTxScheduleInterval:  30,
 				OutboundTxScheduleLookahead: 100,
+			},
+			{
+				ChainId:                     common.BtcRegtestChain().ChainId,
+				ConfirmationCount:           2,
+				ZetaTokenContractAddress:    "",
+				ConnectorContractAddress:    "",
+				Erc20CustodyContractAddress: "",
+				GasPriceTicker:              5,
+				WatchUtxoTicker:             1,
+				InTxTicker:                  1,
+				OutTxTicker:                 2,
+				OutboundTxScheduleInterval:  2,
+				OutboundTxScheduleLookahead: 5,
 			},
 		},
 	}
