--- x/observer/types/core_params_mainnet.go
@@ -1,69 +1,0 @@
-package types
-
-import (
-	"fmt"
-
-	"github.com/coinbase/rosetta-sdk-go/types"
-	"github.com/zeta-chain/zetacore/common"
-)
-
-func GetCoreParams() CoreParamsList {
-	params := CoreParamsList{
-		CoreParams: []*CoreParams{
-			{
-				ChainId:                     common.EthChain().ChainId,
-				ConfirmationCount:           14,
-				ZetaTokenContractAddress:    "",
-				ConnectorContractAddress:    "",
-				Erc20CustodyContractAddress: "",
-				InTxTicker:                  12,
-				OutTxTicker:                 15,
-				WatchUtxoTicker:             0,
-				GasPriceTicker:              30,
-				OutboundTxScheduleInterval:  30,
-				OutboundTxScheduleLookahead: 60,
-			},
-			{
-				ChainId:                     common.BscMainnetChain().ChainId,
-				ConfirmationCount:           14,
-				ZetaTokenContractAddress:    "",
-				ConnectorContractAddress:    "",
-				Erc20CustodyContractAddress: "",
-				InTxTicker:                  5,
-				OutTxTicker:                 15,
-				WatchUtxoTicker:             0,
-				GasPriceTicker:              30,
-				OutboundTxScheduleInterval:  30,
-				OutboundTxScheduleLookahead: 60,
-			},
-			{
-				ChainId:                     common.BtcMainnetChain().ChainId,
-				ConfirmationCount:           2,
-				ZetaTokenContractAddress:    "",
-				ConnectorContractAddress:    "",
-				Erc20CustodyContractAddress: "",
-				WatchUtxoTicker:             30,
-				InTxTicker:                  120,
-				OutTxTicker:                 60,
-				GasPriceTicker:              30,
-				OutboundTxScheduleInterval:  30,
-				OutboundTxScheduleLookahead: 60,
-			},
-		},
-	}
-	chainList := common.ExternalChainList()
-	requiredParams := len(chainList)
-	availableParams := 0
-	for _, chain := range chainList {
-		for _, param := range params.CoreParams {
-			if chain.ChainId == param.ChainId {
-				availableParams++
-			}
-		}
-	}
-	if availableParams != requiredParams {
-		panic(fmt.Sprintf("Core params are not available for all chains , DefaultChains : %s , CoreParams : %s",
-			types.PrettyPrintStruct(chainList), params.String()))
-	}
-	return params
-}
