--- zetaclient/supported_chains.go
@@ -1,10 +1,0 @@
-package zetaclient
-
-import (
-	"github.com/zeta-chain/zetacore/common"
-)
-
-// Modify to update this from the core later
-func GetSupportedChains() []*common.Chain {
-	return common.DefaultChainsList()
-}
